Added separate clients to encapsulate HTTP logic
